requirements.txt
# VisionMap - Python dependencies

torch>=2.1,<3.0
torchvision>=0.16,<1.0
Pillow>=10.0,<11.0

fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
python-multipart>=0.0.6

asyncpg>=0.29,<1.0
pgvector>=0.3,<1.0
